Change layout refactory algorithm
Intermediate layouts cannot be totally removed, since they might be refered in the Java source code.

Hence, it is required to keep then in he refactored layout. The layout refactory algorithm changed so that the new layout contains a root LinearLayout with two LinearLayout child elements, one for Views and the other for non-view elements.