Add feature
An important idea in computing method complexity is to consider the specific methods of Android API that are called by that method. However, due to the use of interfaces and abstract classes, it is possible that those API methods, e.g. database-related methods, are not directly called by a method, but they are called through polymorphism on intrfaces or abstract classes. To consider this issue, in MethodComplexityAnalyzer, when the list of calledMethods of a method is retrieved, it is processed to replace a method with its concrete implementations.

I started to implement this feature. Implementation is completed.

More Android API methods are included in getAPIComplexity.

getMethodComplexty is modified to include IntentComplexity and AsyncComplexity only when the method is not called by ClassComplexityAnalyzer. It means that when computing complexity of a class, the ItentComplexity and AsyncComplexity of its methods are not included, but API complexity is.